+++ packages/@react-native-windows/tester/src/js/examples-win/LegacyTests/DirectManipulationTestPage.tsx
