+++ packages/@react-native-windows/tester/src/js/examples-win/Mouse/MouseExample.windows.js
