+++ packages/@react-native-windows/tester/src/js/examples-win/Picker/PickerWindowsExample.tsx
@@ -6,7 +6,7 @@
 
 import * as React from 'react';
 import {Button, Text, View} from 'react-native';
-import {Picker} from '../../../../Libraries/Components/Picker/PickerWindows';
+import {Picker} from 'react-native-windows';
 
 interface MakesModels {
   name: string;
