+++ packages/@react-native-windows/tester/src/js/examples-win/TransferProperties/TransferPropertiesExample.tsx
