+++ packages/@react-native-windows/tester/src/js/examples/View/ViewExample.windows.js
