--- vnext/src/RNTester.ts
@@ -1,6 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation.
- * Licensed under the MIT License.
- * @format
- */
-import './RNTester/js/RNTesterApp';
